requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langchain-core>=0.2.0
litellm>=1.35.0
anthropic>=0.25.0
pandas>=2.0.0
openpyxl>=3.1.2
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
